refactoring report/loader abstractions
